Added  @MediaType constraint for Mutipart files, extended the functionality of @Min/MaxLength to support MultipartFile size, updated SoletAPI to provide the file size in MemoryFile interface.

+++ SummerMVC_Framework/src/main/java/com/cyecize/summer/areas/validation/constraints/MinLength.java
@@ -11,7 +11,7 @@ import java.lang.annotation.Target;
 @Target(ElementType.FIELD)
 @Constraint(validatedBy = MinLengthConstraint.class)
 public @interface MinLength {
-    int length();
+    long length();
 
     String message() default "String too short";
 }
